Update README.md
poems
=====




A collection of work in progress poems written in markdown. The idea of the repo is to provide me with a way to write and edit poetry in a way that exposes the versions those poems go through. The idea is to open the creative process to technical inspection.




##A Coat




I MADE my song a coat




Covered with embroideries




Out of old mythologies




From heel to throat;




But the fools caught it,




Wore it in the world's eyes




As though they'd wrought it.




Song, let them take it,




For there's more enterprise




In walking naked. 




-William Butler Yeats