Re-working the second paragraph.
The changes here reflect my evolving approach to this repo as either a poem or a set of poems. These current edits are the beginnings of a move away from placing importance on such traditional distinctions and instead looking more intently at the nature of process and its effect on creative output. This tack may be a reflection of recent readings in Benjamin and a re-affirmation of how process informs art in the age of mechanical reproduction.